requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Babel==4.0.0
google-generativeai==0.5.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
APScheduler==3.10.4
Werkzeug==3.0.1